+++ src/test/ExSln2/c99/gen/hal_RefParamsExample.h
@@ -12,4 +12,4 @@ void hal_RefParamsExample_test();
 
 void hal_RefParamsExample_inc(uint8_t * a);
 
-uint8_t hal_RefParamsExample_echo(const uint8_t a);
+uint8_t hal_RefParamsExample_echo(uint8_t const a);