+++ src/test/ExSln2/c99/gen/ts_MemFieldEx.h
@@ -9,7 +9,6 @@
 
 
 
-
 typedef struct ts_MemFieldEx ts_MemFieldEx;
 struct ts_MemFieldEx
 {